Vdrtools & Modular Libs as feature flags (#763)
* Add feature flags to opt-in vdrtools or modular libs implementation
