+++ aries_vcx/src/core/profile/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "modular_libs")]
 pub mod modular_libs_profile;
 pub mod profile;
+#[cfg(feature = "vdrtools")]
 pub mod vdrtools_profile;
