+++ aries_vcx/src/errors/mod.rs
@@ -1,8 +1,11 @@
 pub mod error;
 mod mapping_agency_client;
+#[cfg(feature = "modular_libs")]
 mod mapping_credx;
 mod mapping_diddoc;
+#[cfg(feature = "modular_libs")]
 mod mapping_indyvdr;
 mod mapping_messages;
 mod mapping_others;
+#[cfg(feature = "vdrtools")]
 mod mapping_vdrtools;
