+++ aries_vcx/src/global/settings.rs
@@ -2,7 +2,6 @@ use std::collections::HashMap;
 use std::sync::RwLock;
 
 use crate::errors::error::prelude::*;
-use crate::indy::wallet::IssuerConfig;
 
 pub static CONFIG_POOL_NAME: &str = "pool_name";
 pub static CONFIG_SDK_TO_REMOTE_ROLE: &str = "sdk_to_remote_role";
@@ -185,7 +184,7 @@ pub mod unit_tests {
     }
 }
 
-pub fn init_issuer_config(config: &IssuerConfig) -> VcxResult<()> {
-    set_config_value(CONFIG_INSTITUTION_DID, &config.institution_did)?;
+pub fn init_issuer_config(institution_did: &str) -> VcxResult<()> {
+    set_config_value(CONFIG_INSTITUTION_DID, institution_did)?;
     Ok(())
 }
