+++ aries_vcx/src/plugins/anoncreds/indy_anoncreds.rs
@@ -1,5 +1,3 @@
-use std::sync::Arc;
-
 use async_trait::async_trait;
 use vdrtools::{PoolHandle, WalletHandle};
