+++ aries_vcx/src/plugins/anoncreds/mod.rs
@@ -1,3 +1,5 @@
 pub mod base_anoncreds;
+#[cfg(feature = "modular_libs")]
 pub mod credx_anoncreds;
+#[cfg(feature = "vdrtools")]
 pub mod indy_anoncreds;
