+++ aries_vcx/src/plugins/ledger/mod.rs
@@ -1,3 +1,5 @@
 pub mod base_ledger;
+#[cfg(feature = "vdrtools")]
 pub mod indy_ledger;
+#[cfg(feature = "modular_libs")]
 pub mod indy_vdr_ledger;
