+++ aries_vcx/src/plugins/wallet/mod.rs
@@ -1,3 +1,4 @@
 pub mod agency_client_wallet;
 pub mod base_wallet;
+#[cfg(feature = "vdrtools")]
 pub mod indy_wallet;
