+++ aries_vcx/src/utils/mockdata/profile/mock_anoncreds.rs
@@ -1,9 +1,9 @@
 use async_trait::async_trait;
 
 use crate::errors::error::{AriesVcxError, AriesVcxErrorKind, VcxResult};
+use crate::utils::mockdata::mock_settings::StatusCodeMock;
 use crate::{
     global::settings,
-    indy::utils::LibindyMock,
     plugins::anoncreds::base_anoncreds::BaseAnonCreds,
     utils::{
         self,
@@ -65,8 +65,7 @@ impl BaseAnonCreds for MockAnoncreds {
     }
 
     async fn issuer_create_credential_offer(&self, _cred_def_id: &str) -> VcxResult<String> {
-        let rc = LibindyMock::get_result();
-        if rc != 0 {
+        if StatusCodeMock::get_result() != 0 {
             return Err(AriesVcxError::from_msg(
                 AriesVcxErrorKind::InvalidState,
                 "Mocked error result of issuer_create_credential_offer: issuer_create_credential_offer",
