+++ aries_vcx/src/utils/mockdata/profile/mock_ledger.rs
@@ -1,9 +1,9 @@
 use async_trait::async_trait;
 
 use crate::errors::error::{AriesVcxError, AriesVcxErrorKind, VcxResult};
+use crate::utils::mockdata::mock_settings::StatusCodeMock;
 use crate::{
     common::primitives::revocation_registry::RevocationRegistryDefinition,
-    indy::utils::LibindyMock,
     plugins::ledger::base_ledger::BaseLedger,
     utils::{
         self,
@@ -71,8 +71,7 @@ impl BaseLedger for MockLedger {
     async fn get_cred_def(&self, cred_def_id: &str, submitter_did: Option<&str>) -> VcxResult<String> {
         // TODO - FUTURE - below error is required for tests to pass which require a cred def to not exist (libvcx)
         // ideally we can migrate away from it
-        let rc = LibindyMock::get_result();
-        if rc == 309 {
+        if StatusCodeMock::get_result() == 309 {
             return Err(AriesVcxError::from_msg(
                 AriesVcxErrorKind::LedgerItemNotFound,
                 "Mocked error".to_string(),
